+++ packages/core/src/config/constants.ts
@@ -27,3 +27,6 @@ const { version } = JSON.parse(
 );
 
 export const VERSION = version;
+
+export const ENV_PRODUCTION = 'production';
+export const ENV_DEVELOPMENT = 'development';