+++ packages/core/src/config/env.ts
@@ -12,6 +12,7 @@ import path from 'node:path';
 
 import { parse } from 'dotenv';
 import { type DotenvPopulateInput, expand } from 'dotenv-expand';
+
 import { arraify, normalizePath, tryStatSync } from '../utils/index.js';
 
 export function loadEnv(