+++ packages/core/src/config/index.ts
@@ -1,8 +1,10 @@
 import { createHash } from 'node:crypto';
 import fs from 'node:fs';
+import { readFile } from 'node:fs/promises';
 import { createRequire } from 'node:module';
 import path from 'node:path';
 import { pathToFileURL } from 'node:url';
+
 import fse from 'fs-extra';
 
 import { bindingPath } from '../../binding/index.js';
@@ -23,6 +25,7 @@ import {
   colors,
   isArray,
   isEmptyObject,
+  isNodeEnv,
   isObject,
   isWindows,
   normalizePath
@@ -43,25 +46,33 @@ import { normalizePersistentCache } from './normalize-config/normalize-persisten
 import { parseUserConfig } from './schema.js';
 
 import { externalAdapter } from '../plugin/js/external-adapter.js';
+import { ViteModuleGraphAdapter } from '../plugin/js/vite-server-adapter.js';
 import { convertErrorMessage } from '../utils/error.js';
 import { resolveHostname } from '../utils/http.js';
 import merge from '../utils/merge.js';
 import {
   CUSTOM_KEYS,
   DEFAULT_CONFIG_NAMES,
+  ENV_DEVELOPMENT,
+  ENV_PRODUCTION,
   FARM_DEFAULT_NAMESPACE
 } from './constants.js';
 import { mergeConfig, mergeFarmCliConfig } from './mergeConfig.js';
+
 import { normalizeCss } from './normalize-config/normalize-css.js';
 import { normalizeExternal } from './normalize-config/normalize-external.js';
-import normalizePartialBundling from './normalize-config/normalize-partial-bundling.js';
+import { normalizePartialBundling } from './normalize-config/normalize-partial-bundling.js';
 import { normalizeResolve } from './normalize-config/normalize-resolve.js';
 
-import type { OutputConfig } from '../types/binding.js';
 import type {
   ConfigEnv,
+  DefaultOptionsType,
+  EnvResult,
   FarmCliOptions,
+  Format,
   HmrOptions,
+  ModuleContext,
+  ModuleNode,
   NormalizedServerConfig,
   ResolvedCompilation,
   ResolvedUserConfig,
@@ -86,6 +97,8 @@ export function defineFarmConfig(config: UserConfigExport): UserConfigExport;
 export function defineFarmConfig(config: UserConfigExport): UserConfigExport {
   return config;
 }
+
+type UserConfigPromise = Promise<UserConfig | undefined>;
 
 const COMMANDS = {
   START: 'start',
@@ -210,6 +223,8 @@ async function handleResolveConfig(
     root: resolvedUserConfig.compilation.root,
     jsPlugins: sortFarmJsPlugins,
     rustPlugins: rustPlugins,
+    command,
+    isProduction: resolvedUserConfig.compilation.mode === ENV_PRODUCTION,
     transformInlineConfig
   });
 
@@ -289,8 +304,8 @@ export async function normalizeUserCompilationConfig(
     compilation
   );
 
-  const isProduction = mode === 'production';
-  const isDevelopment = mode === 'development';
+  const isProduction = mode === ENV_PRODUCTION;
+  const isDevelopment = mode === ENV_DEVELOPMENT;
   resolvedCompilation.mode = resolvedCompilation.mode ?? mode;
 
   resolvedCompilation.coreLibPath = bindingPath;
@@ -333,61 +348,44 @@ export async function normalizeUserCompilationConfig(
     // for node target, we should not define process.env.NODE_ENV
     resolvedCompilation.output?.targetEnv === 'node'
       ? {}
-      : Object.keys(resolvedUserConfig.env || {}).reduce((env: any, key) => {
-          env[`$__farm_regex:(global(This)?\\.)?process\\.env\\.${key}`] =
-            JSON.stringify(resolvedUserConfig.env[key]);
-          return env;
-        }, {})
+      : Object.keys(resolvedUserConfig.env || {}).reduce<EnvResult>(
+          (env, key) => {
+            env[`$__farm_regex:(global(This)?\\.)?process\\.env\\.${key}`] =
+              JSON.stringify(resolvedUserConfig.env[key]);
+            return env;
+          },
+          {} as EnvResult
+        )
   );
 
   const require = createRequire(import.meta.url);
   const hmrClientPluginPath = require.resolve('@farmfe/runtime-plugin-hmr');
-  const ImportMetaPluginPath = require.resolve(
+  const importMetaPluginPath = require.resolve(
     '@farmfe/runtime-plugin-import-meta'
   );
 
-  if (!resolvedCompilation.runtime) {
-    resolvedCompilation.runtime = {
-      path: require.resolve('@farmfe/runtime'),
-      plugins: []
-    };
-  }
+  resolvedCompilation.runtime = {
+    path:
+      resolvedCompilation.runtime?.path ?? require.resolve('@farmfe/runtime'),
+    swcHelpersPath:
+      resolvedCompilation.runtime?.swcHelpersPath ??
+      path.dirname(require.resolve('@swc/helpers/package.json')),
+    plugins: resolvedCompilation.runtime?.plugins ?? [],
+    namespace: resolvedCompilation.runtime?.namespace
+  };
 
-  if (!resolvedCompilation.runtime.path) {
-    resolvedCompilation.runtime.path = require.resolve('@farmfe/runtime');
-  }
-
-  if (!resolvedCompilation.runtime.swcHelpersPath) {
-    resolvedCompilation.runtime.swcHelpersPath = path.dirname(
-      require.resolve('@swc/helpers/package.json')
-    );
-  }
-
-  if (!resolvedCompilation.runtime.plugins) {
-    resolvedCompilation.runtime.plugins = [];
-  } else {
-    const resolvePluginPath = (plugin: any) => {
+  resolvedCompilation.runtime.plugins = resolvedCompilation.runtime.plugins.map(
+    (plugin) => {
       if (path.isAbsolute(plugin)) return plugin;
       return plugin.startsWith('.')
         ? path.resolve(resolvedRootPath, plugin)
         : require.resolve(plugin);
-    };
-    // make sure all plugin paths are absolute
-    resolvedCompilation.runtime.plugins =
-      resolvedCompilation.runtime.plugins.map(resolvePluginPath);
-  }
-  // set namespace to package.json name field's hash
-  if (!resolvedCompilation.runtime.namespace) {
-    // read package.json name field
-    const packageJsonPath = path.resolve(resolvedRootPath, 'package.json');
-    const packageJsonExists = fse.existsSync(packageJsonPath);
-    const namespaceName = packageJsonExists
-      ? JSON.parse(fse.readFileSync(packageJsonPath, 'utf-8')).name ||
-        FARM_DEFAULT_NAMESPACE
-      : FARM_DEFAULT_NAMESPACE;
+    }
+  );
 
+  if (!resolvedCompilation.runtime.namespace) {
     resolvedCompilation.runtime.namespace = createHash('md5')
-      .update(namespaceName)
+      .update(getNamespaceName(resolvedRootPath))
       .digest('hex');
   }
 
@@ -400,16 +398,9 @@ export async function normalizeUserCompilationConfig(
   resolvedCompilation.mode ??= mode;
 
   setProcessEnv(resolvedCompilation.mode);
-
-  // TODO add targetEnv `lib-browser` and `lib-node` support
-  const is_entry_html =
-    !resolvedCompilation.input ||
-    Object.values(resolvedCompilation.input).some(
-      (value) => value && value.endsWith('.html')
-    );
-
+  const isNode = isNodeEnv(resolvedCompilation.output.targetEnv);
   if (
-    resolvedCompilation.output.targetEnv !== 'node' &&
+    !isNode &&
     isArray(resolvedCompilation.runtime.plugins) &&
     resolvedUserConfig.server?.hmr &&
     !resolvedCompilation.runtime.plugins.includes(hmrClientPluginPath)
@@ -440,9 +431,9 @@ export async function normalizeUserCompilationConfig(
 
   if (
     isArray(resolvedCompilation.runtime.plugins) &&
-    !resolvedCompilation.runtime.plugins.includes(ImportMetaPluginPath)
+    !resolvedCompilation.runtime.plugins.includes(importMetaPluginPath)
   ) {
-    resolvedCompilation.runtime.plugins.push(ImportMetaPluginPath);
+    resolvedCompilation.runtime.plugins.push(importMetaPluginPath);
   }
 
   // we should not deep merge compilation.input
@@ -567,7 +558,7 @@ export const DEFAULT_DEV_SERVER_OPTIONS: NormalizedServerConfig = {
     name: 'localhost',
     host: undefined
   },
-  host: true,
+  host: 'localhost',
   proxy: undefined,
   hmr: DEFAULT_HMR_OPTIONS,
   middlewareMode: false,
@@ -577,6 +568,7 @@ export const DEFAULT_DEV_SERVER_OPTIONS: NormalizedServerConfig = {
   middlewares: [],
   appType: 'spa',
   writeToDisk: false,
+  origin: '',
   preview: {
     host: 'localhost',
     headers: {},
@@ -613,8 +605,6 @@ export const DEFAULT_COMPILATION_OPTIONS: Partial<ResolvedCompilation> = {
     ]
   }
 };
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
 
 function tryHttpsAsFileRead(value: unknown): string | Buffer | unknown {
   if (typeof value === 'string') {
@@ -663,7 +653,6 @@ export function normalizeDevServerConfig(
   }) as NormalizedServerConfig;
 }
 
-type Format = Exclude<OutputConfig['format'], undefined>;
 const formatFromExt: Record<string, Format> = {
   cjs: 'cjs',
   mjs: 'esm',
@@ -682,7 +671,7 @@ export async function readConfigFile(
   configFilePath: string,
   configEnv: ConfigEnv,
   mode: CompilationMode = 'development'
-): Promise<UserConfig | undefined> {
+): UserConfigPromise {
   if (!fse.existsSync(configFilePath)) return;
 
   const format = getFormat(configFilePath);
@@ -818,7 +807,7 @@ export async function loadConfigFile(
     const errorMessage = convertErrorMessage(error);
     const stackTrace =
       error.code === 'GenericFailure' ? '' : `\n${error.stack}`;
-    if (inlineOptions.mode === 'production') {
+    if (inlineOptions.mode === ENV_PRODUCTION) {
       throw new Error(
         `Failed to load farm config file: ${errorMessage} \n${stackTrace}`
       );
@@ -842,7 +831,7 @@ export async function checkCompilationInputValue(
   const { compilation } = userConfig;
   const targetEnv = compilation?.output?.targetEnv;
   const inputValue = Object.values(compilation?.input).filter(Boolean);
-  const isTargetNode = targetEnv === 'node';
+  const isTargetNode = isNodeEnv(targetEnv);
   const defaultHtmlPath = './index.html';
   let inputIndexConfig: {
     index?: string;
@@ -890,7 +879,6 @@ export async function checkCompilationInputValue(
         `Build failed due to errors: Can not resolve ${
           isTargetNode ? 'index.js or index.ts' : 'index.html'
         }  from ${userConfig.root}. \n${errorMessage}`
-        // { exit: true }
       );
     }
   }
@@ -925,8 +913,9 @@ export async function resolvePlugins(
 ) {
   const { jsPlugins: rawJsPlugins, rustPlugins } =
     await resolveFarmPlugins(userConfig);
-  const jsPlugins = await resolveAndFilterAsyncPlugins(rawJsPlugins);
-
+  const jsPlugins = (await resolveAndFilterAsyncPlugins(rawJsPlugins)).map(
+    wrapPluginUpdateModules
+  );
   const vitePlugins = (userConfig?.vitePlugins ?? []).filter(Boolean);
 
   const vitePluginAdapters = vitePlugins.length
@@ -941,7 +930,7 @@ export async function resolvePlugins(
   };
 }
 
-export async function resolveDefaultUserConfig(options: any) {
+export async function resolveDefaultUserConfig(options: DefaultOptionsType) {
   const defaultConfig: UserConfig = createDefaultConfig(options);
 
   const resolvedUserConfig: ResolvedUserConfig = await resolveUserConfig(
@@ -951,7 +940,7 @@ export async function resolveDefaultUserConfig(options: any) {
 
   const normalizedConfig = await normalizeUserCompilationConfig(
     resolvedUserConfig,
-    'development'
+    options.mode
   );
 
   return normalizedConfig;
@@ -999,8 +988,8 @@ export async function resolveUserConfig(
     // TODO publicPath rewrite to BASE_URL
     BASE_URL: userConfig.compilation.output.publicPath ?? '/',
     mode: userConfig.mode,
-    DEV: userConfig.compilation.mode === 'development',
-    PROD: userConfig.compilation.mode === 'production'
+    DEV: userConfig.compilation.mode === ENV_DEVELOPMENT,
+    PROD: userConfig.compilation.mode === ENV_PRODUCTION
   };
 
   resolvedUserConfig.publicDir = normalizePublicDir(
@@ -1011,7 +1000,7 @@ export async function resolveUserConfig(
   return resolvedUserConfig;
 }
 
-export function createDefaultConfig(options: any): UserConfig {
+export function createDefaultConfig(options: DefaultOptionsType): UserConfig {
   const { inlineOptions, mode, format, outputPath, fileName, configFilePath } =
     options;
 
@@ -1115,4 +1104,51 @@ async function setLazyCompilationDefine(
       resolvedUserConfig.server.protocol || 'http'
     }://${hostname.host || 'localhost'}:${resolvedUserConfig.server.port}`
   };
+}
+
+function getNamespaceName(rootPath: string) {
+  const packageJsonPath = path.resolve(rootPath, 'package.json');
+  if (fse.existsSync(packageJsonPath)) {
+    const { name } = JSON.parse(fse.readFileSync(packageJsonPath, 'utf-8'));
+    return name || FARM_DEFAULT_NAMESPACE;
+  }
+  return FARM_DEFAULT_NAMESPACE;
+}
+
+function wrapPluginUpdateModules(plugin: JsPlugin): JsPlugin {
+  if (!plugin.updateModules?.executor) {
+    return plugin;
+  }
+  const originalExecutor = plugin.updateModules.executor;
+  const moduleGraph = new ViteModuleGraphAdapter(plugin.name);
+
+  plugin.updateModules.executor = async ({ paths }, ctx) => {
+    moduleGraph.context = ctx;
+    for (const [file, type] of paths) {
+      const mods = moduleGraph.getModulesByFile(
+        file
+      ) as unknown as ModuleNode[];
+
+      const filename = normalizePath(file);
+      const moduleContext: ModuleContext = {
+        file: filename,
+        timestamp: Date.now(),
+        type,
+        paths,
+        modules: (mods ?? []).map(
+          (m) =>
+            ({
+              ...m,
+              id: normalizePath(m.id),
+              file: normalizePath(m.file)
+            }) as ModuleNode
+        ),
+        read: function (): string | Promise<string> {
+          return readFile(file, 'utf-8');
+        }
+      };
+      return originalExecutor.call(plugin, moduleContext);
+    }
+  };
+  return plugin;
 }