+++ packages/core/src/plugin/js/index.ts
@@ -177,6 +177,12 @@ export function convertPluginVite(plugin: JsPlugin): void {
     }
   }
 
+  if (plugin.processModule) {
+    plugin.processModule.filters ??= {};
+    plugin.processModule.filters.moduleTypes ??= [];
+    plugin.processModule.filters.resolvedPaths ??= [];
+  }
+
   if (plugin.renderResourcePot) {
     plugin.renderResourcePot.filters ??= {};