+++ packages/core/src/server/middlewares/index.ts
@@ -3,7 +3,6 @@ export * from './lazyCompilation.js';
 export * from './notFound.js';
 export * from './proxy.js';
 export * from './resource.js';
-export * from './publicResource.js';
 export * from './publicPath.js';
 export * from './hmrPing.js';
-export * from './adaptorVite.js';
+export * from './static.js';