+++ crates/plugin_script/CHANGELOG.md
@@ -6,6 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.12](https://github.com/farm-fe/farm/compare/farmfe_plugin_script-v0.0.11...farmfe_plugin_script-v0.0.12) - 2024-12-09
+
+### Added
+
+- v1.5.0 (#1987)
+
+### Fixed
+
+- new import meta url runtime & merge electron preload output files (#1984)
+- *(runtime)* async cache (#1902)
+
 ## [0.0.11](https://github.com/farm-fe/farm/compare/farmfe_plugin_script-v0.0.10...farmfe_plugin_script-v0.0.11) - 2024-10-31
 
 ### Fixed