+++ packages/core/CHANGELOG.md
@@ -5,6 +5,15 @@
 ### Major Changes
 
 - b327dd1: collect htmlFallbackMiddleware
+## 1.6.0
+
+### Minor Changes
+
+- 6b849123: js plugin supports process module hook
+
+### Patch Changes
+
+- b235a919: feat: support invalidateModule
 
 ## 1.5.0